app/requirements.txt
streamlit==1.23.1
Pillow==9.5.0
torchvision==0.15.2
torch==2.0.1